第九模块-前后端分离开发v2.0/第1部分：drf/dig/requirements.txt
django==3.2.6
django-filter==2.4.0
django-redis==5.0.0
djangorestframework==3.12.4
